Added some comments about bugs that currently have no trivial solution right now. Correct some compiler warnings. Correct a bug where iterator death listeners weren't working. Some minor refactorings in preparation for song iterator unit tests (to be checked in a little later).

+++ src/scicore/sci_memory.c
@@ -155,7 +155,7 @@ _SCI_STRNDUP(const char *src, size_t length, char *file, int line, char *funct)
 {
 	void *res;
 	char *strres;
-	int rlen = MIN(strlen(src), length) + 1;
+	int rlen = (int)MIN(strlen(src), length) + 1;
 #ifdef MALLOC_DEBUG
 	INFO_MEMORY("_SCI_STRNDUP()", 0, file, line, funct);
 #endif
